Update app version to 1.1.0 and fix update functionality
This commit updates the app version to 1.1.0 and fixes the update functionality. It adds a new feature to check for available updates and prompts the user to update the app if a new version is available. The code also includes a test message for debugging purposes.

Changes:
- Updated app version to 1.1.0 in app.json
- Added code to check for available updates and prompt the user to update
- Fixed the update functionality to fetch and reload the app after updating
- Added a test message for debugging purposes

Refactor code and update AndroidManifest.xml

This commit refactors the code and updates the AndroidManifest.xml file. It removes unnecessary code related to TV features and adds a new meta-data tag for the app name.

Changes:
- Removed unused code related to TV features in AndroidManifest.xml
- Added a new meta-data tag for the app name in AndroidManifest.xml

Update package-lock.json and package.json

This commit updates the package-lock.json and package.json files. It updates the expo-updates dependency to version ^0.24.13.

Changes:
- Updated expo-updates dependency to version ^0.24.13 in package-lock.json and package.json

Update Navbar component

This commit updates the Navbar component. It changes the version number displayed in the dropdown menu to v1.0.2.

Changes:
- Updated the version number displayed in the dropdown menu to v1.0.2 in Navbar.tsx